+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version history for `cardano-ledger-api`
 
+## 1.2.0.0
+
+* Add support for Plutus V3
+
 ## 1.1.0.0
 
 * Add `Cardano.Ledger.Api.Governance` that exposes all of the governance related types.